+++ package.json
@@ -21,8 +21,8 @@
   "scripts": {
     "build": "npm run build:clean && npm run build:module && npm run build:commonjs && npm run build:types",
     "build:clean": "rm -rf dist",
-    "build:module": "  tsc --module es2015 && mv dist/ts-out/*.js dist && npm run build:module:import",
-    "build:module:import": "  sed -i'.bak' -e 's/\\.\\/linkedlist/.\\/linkedlist.js/' -e 's/\\.\\/taskscheduler/.\\/taskscheduler.js/'  dist/asynciterator.js  && rm dist/*.bak",
+    "build:module": "tsc && mv dist/ts-out/*.js dist && npm run build:module:import",
+    "build:module:import": "sed -i'.bak' -e 's/\\.\\/linkedlist/.\\/linkedlist.js/' -e 's/\\.\\/taskscheduler/.\\/taskscheduler.js/'  dist/asynciterator.js  && rm dist/*.bak",
     "build:commonjs": "tsc --module commonjs && ./.change-extension cjs dist/ts-out/*.js && mv dist/ts-out/*.cjs dist && npm run build:commonjs:import",
     "build:commonjs:import": "sed -i'.bak' -e 's/\\.\\/linkedlist/.\\/linkedlist.cjs/' -e 's/\\.\\/taskscheduler/.\\/taskscheduler.cjs/' dist/asynciterator.cjs && rm dist/*.bak",
     "build:types": "tsc -d && rm dist/ts-out/*.js && mv dist/ts-out/*.d.ts dist",
