+++ test/SimpleTransformIterator-test.js
@@ -1299,107 +1299,6 @@ describe('SimpleTransformIterator', () => {
     });
   });
 
-  describe('The AsyncIterator#skip function', () => {
-    it('should be a function', () => {
-      expect(AsyncIterator.prototype.skip).to.be.a('function');
-    });
-
-    describe('when called on an iterator', () => {
-      let iterator, result;
-      before(() => {
-        iterator = new ArrayIterator(['a', 'b', 'c', 'd', 'e']);
-        result = iterator.skip(2);
-      });
-
-      describe('the return value', () => {
-        const items = [];
-        before(done => {
-          result.on('data', item => { items.push(item); });
-          result.on('end', done);
-        });
-
-        it('should skip the given number of items', () => {
-          items.should.deep.equal(['c', 'd', 'e']);
-        });
-      });
-    });
-  });
-
-  describe('The AsyncIterator#take function', () => {
-    it('should be a function', () => {
-      expect(AsyncIterator.prototype.take).to.be.a('function');
-    });
-
-    describe('when called on an iterator', () => {
-      let iterator, result;
-      before(() => {
-        iterator = new ArrayIterator(['a', 'b', 'c', 'd', 'e']);
-        result = iterator.take(3);
-      });
-
-      describe('the return value', () => {
-        const items = [];
-        before(done => {
-          result.on('data', item => { items.push(item); });
-          result.on('end', done);
-        });
-
-        it('should take the given number of items', () => {
-          items.should.deep.equal(['a', 'b', 'c']);
-        });
-      });
-    });
-  });
-
-  describe('The AsyncIterator#range function', () => {
-    it('should be a function', () => {
-      expect(AsyncIterator.prototype.range).to.be.a('function');
-    });
-
-    describe('when called on an iterator', () => {
-      let iterator, result;
-      before(() => {
-        iterator = new IntegerIterator();
-        result = iterator.range(20, 29);
-      });
-
-      describe('the return value', () => {
-        const items = [];
-        before(done => {
-          result.on('data', item => { items.push(item); });
-          result.on('end', done);
-        });
-
-        it('should contain the indicated range', () => {
-          items.should.have.length(10);
-          items[0].should.equal(20);
-          items[9].should.equal(29);
-        });
-      });
-    });
-
-    describe('when called on an iterator with an inverse range', () => {
-      let iterator, result;
-      before(() => {
-        iterator = new IntegerIterator();
-        sinon.spy(iterator, 'read');
-      });
-
-      describe('the return value', () => {
-        const items = [];
-        before(done => {
-          result = iterator.range(30, 20);
-          result.on('data', item => { items.push(item); });
-          result.on('end', done);
-        });
-
-        it('should be empty', () => {
-          items.should.be.empty;
-        });
-      });
-    });
-  });
-
   describe('The AsyncIterator#transform function', () => {
     it('should be a function', () => {
       expect(AsyncIterator.prototype.transform).to.be.a('function');
